requirements.txt
flake8==3.7.1
pycodestyle==2.5.0
PyYAML==5.1.1
nose==1.3.7
coverage==4.5.3   
